lambda/IoTCertificateValidator/requirements.txt
pyOpenSSL 